CMS-322: Move functionality to a hook
